mod file location and mod login style
